Add buttons to retropad.

+++ input/drivers_joypad/linuxraw_joypad.c
@@ -345,13 +345,13 @@ static int16_t linuxraw_joypad_axis(unsigned port, uint32_t joyaxis)
    return linuxraw_joypad_axis_state(pad, port, joyaxis);
 }
 
-static int16_t linuxraw_joypad_state(
+static int32_t linuxraw_joypad_state(
       rarch_joypad_info_t *joypad_info,
       const struct retro_keybind *binds,
       unsigned port)
 {
    unsigned i;
-   int16_t ret                          = 0;
+   int32_t ret                          = 0;
    uint16_t port_idx                    = joypad_info->joy_idx;
    const struct linuxraw_joypad    *pad = (const struct linuxraw_joypad*)
       &linuxraw_pads[port_idx];
