+++ input/drivers_joypad/sdl_joypad.c
@@ -412,13 +412,13 @@ static int16_t sdl_joypad_axis(unsigned port, uint32_t joyaxis)
    return sdl_joypad_axis_state(pad, port, joyaxis);
 }
 
-static int16_t sdl_joypad_state(
+static int32_t sdl_joypad_state(
       rarch_joypad_info_t *joypad_info,
       const struct retro_keybind *binds,
       unsigned port)
 {
    unsigned i;
-   int16_t ret                          = 0;
+   int32_t ret                          = 0;
    uint16_t port_idx                    = joypad_info->joy_idx;
    sdl_joypad_t *pad                    = (sdl_joypad_t*)&sdl_pads[port_idx];
